requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
akshare==1.17.88
attrs==25.4.0
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
bs4==0.0.2
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
deap==1.4.3
decorator==5.2.1
et_xmlfile==2.0.0
fastjsonschema==2.21.2
Flask==3.1.2
Flask-SocketIO==5.6.0
frozenlist==1.8.0
h11==0.16.0
html5lib==1.1
hypothesis==6.151.4
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpath==0.82.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.3
mini-racer==0.14.0
multidict==6.7.1
narwhals==2.15.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.4.1
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
peewee==3.19.0
platformdirs==4.5.1
plotly==6.5.2
pluggy==1.6.0
propcache==0.4.1
psutil==7.2.1
pycparser==3.0
Pygments==2.19.2
psycopg2-binary==2.9.11
pyproject_hooks==1.2.0
pyqtgraph==0.14.0
PySide6==6.8.2.1
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-socketio==5.16.0
pytz==2025.2
pyzmq==27.1.0
tomli==2.2.1; python_version < '3.11'
QDarkStyle==3.2.3
QtPy==2.4.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
shiboken6==6.8.2.1
simple-websocket==1.1.0
simplejson==3.20.2
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8.3
TA-Lib==0.6.8
tabulate==0.9.0
tqdm==4.67.1
traitlets==5.14.3
tushare==1.4.24
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vnpy==4.2.0
vnpy_ctp==6.7.11.1
vnpy_datarecorder==1.1.1
vnpy_postgresql==1.1.1
vnpy_portfoliostrategy==1.2.2
vnpy_sopt==3.7.1.2
vnpy_spreadtrading==1.3.1
vnpy_tushare==1.4.21.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.5
win32_setctime==1.2.0
wsproto==1.3.2
xlrd==2.0.2
yarl==1.22.0
